Knock Saving format versions down to 0 to indicate indev/not ready

+++ Saving/CreationSaveData.cs
@@ -5,7 +5,7 @@ namespace Root.Saving;
 
 public sealed class CreationSaveData
 {
-	public ushort Version { get; init; } = 1;
+	public byte Version { get; init; } = 1;
 	public DateTimeOffset UtcCreatedAt { get; init; } = GTimeProvider.GetUtcNow();
 
 #pragma warning disable MA0016
